requirements.txt
pika==1.3.2
requests==2.31.0
# Version 3.7.0+ requise pour un support stable de Python 3.12
snowflake-connector-python>=3.7.0
dbt-core>=1.7.11
dbt-snowflake>=1.7.2
python-dotenv==1.0.0
dbt-snowflake